backend/pyproject.toml
[project]
name = "backend"
version = "0.4.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
